add exist check, async get ugc list.

+++ main.py
@@ -1,33 +1,37 @@
 import time
 import asyncio
 import httpx
-from tqdm.auto import tqdm
 import sys
 import re
 import os
+import tqdm.asyncio
+# from pprint import pprint
 
 path = r'./music/'
 if not os.path.exists(path):
     os.makedirs(path)
 
 
-async def download_music(client: httpx.AsyncClient, shareid: str, title: str):
+async def download_music(client: httpx.AsyncClient, shareid: str, title: str, nickname:str):
     data_url = "http://cgi.kg.qq.com/fcgi-bin/fcg_get_play_url?shareid=" + str(shareid)
 
-    file = path + title + '.m4a'
+    file = path + title + '-' + str(nickname) + '.m4a'
+    if os.path.exists(file):
+        print(f"File {file} already exists")
+        return
 
     async with client.stream("GET", data_url, timeout = 5) as response:
         with open(file, 'wb') as f:
-            with tqdm(
+            with tqdm.asyncio.tqdm(
                 unit='B', unit_scale=True, unit_divisor=1024, miniters=1,
-                desc=title, total=int(response.headers.get('content-length', 0))
+                desc=title + '-' + nickname, total=int(response.headers.get('content-length', 0))
             ) as pbar:
                 async for chunk in response.aiter_bytes(chunk_size=4096):
                     f.write(chunk)
                     pbar.update(len(chunk))
         
 
-async def get_music_url_list(client: httpx.AsyncClient, uid: str):
+async def get_music_data(client: httpx.AsyncClient, uid: str):
     start_count = 1
     ugc_list_all = []
 
@@ -35,6 +39,14 @@ async def get_music_url_list(client: httpx.AsyncClient, uid: str):
         "user-agent": "Mozilla/5.0 (iPhone; CPU iPhone OS 16_6 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/16.6 Mobile/15E148 Safari/604.1",
         "referer": "https://static-play.kg.qq.com/node/personal_v2/?uid=" + str(uid),
     }
+
+    ts_ms = int(time.time() * 1000)
+    url = "https://node.kg.qq.com/fcgi-bin/kg_ugc_get_homepage?outCharset=utf-8&from=1&nocache={}&format=json&type=get_uinfo&start={}&num=10&share_uid={uid}&g_tk=1164660242&g_tk_openkey=1164660242".format(ts_ms, start_count, uid=uid)
+
+    resp = await client.get(url, headers=headers)
+    data = resp.json()["data"]
+    ugc_total_count = data["ugc_total_count"]
+    nickname = data["nickname"]
 
     async def _get_url_list(start_count):
         ts_ms = int(time.time() * 1000)
@@ -44,29 +56,33 @@ async def get_music_url_list(client: httpx.AsyncClient, uid: str):
 
         resp = resp.json()
         ugc_list = resp["data"]["ugclist"]
+        ugc_list_all.extend(ugc_list)
         return ugc_list
 
 
-    while True:
-        ugc_list = await _get_url_list(start_count)
-        ugc_list_all.extend(ugc_list)
-        if len(ugc_list) < 10:
-            break
-        start_count += 1
+    get_count = ugc_total_count // 10
+    if ugc_total_count % 10 !=0 :
+        get_count = get_count + 1
 
+    tasks = []
+    for i in range(1, get_count + 1):
+        tasks.append(_get_url_list(i))
 
-    return ugc_list_all
+    await asyncio.gather(*tasks)
+
+    return ugc_list_all, nickname
 
 
 async def main(uid):
      async with httpx.AsyncClient(follow_redirects=True) as client:
-        ugclist = await get_music_url_list(client, uid)
+        ugclist, nickname = await get_music_data(client, uid)
+        print(f"Found {len(ugclist)} songs by {nickname}")
         tasks = []
         for i in range(len(ugclist)):
             ugc = ugclist[i]
             title = ugc["title"]
             shareid = ugc["shareid"]
-            tasks.append(download_music(client, shareid, title))
+            tasks.append(download_music(client, shareid, title, nickname))
             if i%10 == 0:
                 await asyncio.gather(*tasks)
                 tasks = []
